+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2020-04-19 [0.0.2](https://github.com/rahibbert/swi-chrome-extension/pull/6)
+
+- Added Dynamic quote + author + username to the new tab screen
+
 # 2020-04-16 [0.0.1](https://github.com/rahibbert/swi-chrome-extension/pull/5)
 
 - Added changelog feature